# echo-backend/src/main/java/com/theskysid/echobackend/model/ChatMessage.java
package com.theskysid.echobackend.model;

import jakarta.persistence.*;
import lombok.Data;
import java.time.LocalDateTime;

@Data
@Entity
@Table(name = "chat_messages")
public class ChatMessage {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String content;

    private String sender;

    private String recepient;

    private String color;

    @Column(nullable = false)
    private LocalDateTime timestamp;

    @Enumerated(EnumType.STRING)
    private MessageType type;

    public enum MessageType {
       CHAT, PRIVATE_MESSAGE, JOIN, LEAVE, TYPING
    }
}